read.me
Snowhub is a great theme but it has so many colors that it feels like a kid at one point. Although I agreed to all of those colors while developing the theme, after using it for so long I felt it was time to change it.

Snowhub is basically a short version of Snowflat.

If you like the theme, leave your word for it (it helps keep me motivated). I think I will revise this theme in the future too so I would be glad if anyone would like to join it.